server/requirements.txt
annotated-doc
annotated-types
anyio
bcrypt
certifi
cffi
charset-normalizer
click
colorama
cryptography
distro
dnspython
ecdsa
email-validator
eval_type_backport
fastapi
googleapis-common-protos
greenlet
groq
h11
httpcore
httpx
idna
importlib_metadata
invoke
jiter
mistralai
numpy
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
pandas
passlib
prettytable
protobuf
psycopg2-binary
pgvector
pyasn1
pycparser
pydantic
pydantic_core
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
requests
rsa
six
sniffio
SQLAlchemy
starlette
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
zipp